# Try to merge optimizations in tx-buffer