ns-3.30 has a different implementation of `tcp-tx-buffer` compared to the one in ns-3.27. Specifically, 3.30 doesn't have `RateSample`, a pointer to `tcp-socket-base`, and a different implementation of `GetTransmittedSegment()`.

In this tx-buffer, it has all the new implementation in 3.30, while keeping 'RateSample' and other variables from 3.27. It does compile and run, but the results are different - BBR has really low throughput during the simulation.

Before: 
  BBR - avg tput: 1.10069 Mb/s
  Other - avg tput: 1.01866 Mb/s
  Other - avg tput: 0.261758 Mb/s
  Other - avg tput: 0.261758 Mb/s
  Other - avg tput: 1.82414 Mb/s
  Other - avg tput: 1.50983 Mb/s
  Other - avg tput: 1.31409 Mb/s
  Other - avg tput: 1.57165 Mb/s
  Other - avg tput: 0.606648 Mb/s
Jain's fairness index: 0.793337

After:
  BBR - avg tput: 0.285826 Mb/s
  Other - avg tput: 2.62288 Mb/s
  Other - avg tput: 1.45629 Mb/s
  Other - avg tput: 1.44944 Mb/s
  Other - avg tput: 1.15015 Mb/s
  Other - avg tput: 1.10499 Mb/s
  Other - avg tput: 0.511483 Mb/s
  Other - avg tput: 0.511483 Mb/s
  Other - avg tput: 0.511483 Mb/s
Jain's fairness index: 0.706237